Translate the table declarations to use pandoc table management
This change allows to get rid of the hack made on the generated latex.
Moreover, writers and translators can now use markdown typesetting inside
the table and it is be correctly converted.

One downside is that to be processed correctly, the table must be in form of
a block made of lines starting with a tab and with tab as a column separator
ie matching the regex /^\t[^t^n]+\t[^\t\n]+$/. Otherwise it is not transformed
to table.